examples/templates/cdk/launchtemplate/requirements.txt
aws_cdk.core
aws_cdk.aws_iam
aws_cdk.aws_ec2